requirements.txt
psycopg2-binary==2.8.6
channels-redis
tzdata
django-environ
-e ./donation-tracker